Fixed final warning (see last commit), which really was doing something wrong. It only affected humans who would have been born in the last year before the start of warmup, so probably most differences are random.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@1220 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/Population.cpp
@@ -149,7 +149,7 @@ void Population::createInitialHumans ()
 	    // Optimisation: don't create humans who would die before end of
 	    // one-life-span init phase. Transmission models require all humans
 	    // during last year of init; humans bound-to-die before that are unneeded:
-	    if( -iage >= -Global::intervalsPerYear )
+	    if( -iage >= -((int)Global::intervalsPerYear) )
 		newHuman (-iage);
 	    ++cumulativePop;
 	}
